Change the folder structure to use an Http parent folder.
Update integrations.json version (and fix version numbers broken in
https://github.com/DataDog/dd-trace-dotnet/commit/d72b04cb694fd6c2141a59de0baa59568cd8ad49)
