+++ src/Datadog.Trace.ClrProfiler.Managed/AutoInstrumentation/Http/HttpClientHandler/HttpClientHandlerSyncIntegration.cs
@@ -3,7 +3,7 @@ using System.Threading;
 using Datadog.Trace.ClrProfiler.CallTarget;
 using Datadog.Trace.Configuration;
 
-namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.HttpClientHandler
+namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.Http.HttpClientHandler
 {
     /// <summary>
     /// System.Net.Http.HttpClientHandler calltarget instrumentation
@@ -11,13 +11,13 @@ namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.HttpClientHandler
     [InstrumentMethod(
         AssemblyName = "System.Net.Http",
         TypeName = "System.Net.Http.HttpClientHandler",
-        MethodName = "SendAsync",
-        ReturnTypeName = ClrNames.HttpResponseMessageTask,
+        MethodName = "Send",
+        ReturnTypeName = ClrNames.HttpResponseMessage,
         ParameterTypeNames = new[] { ClrNames.HttpRequestMessage, ClrNames.CancellationToken },
-        MinimumVersion = "4.0.0",
+        MinimumVersion = "5.0.0",
         MaximumVersion = "5.*.*",
         IntegrationName = IntegrationName)]
-    public class HttpClientHandlerIntegration
+    public class HttpClientHandlerSyncIntegration
     {
         private const string IntegrationName = nameof(IntegrationIds.HttpMessageHandler);
 
@@ -37,19 +37,20 @@ namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.HttpClientHandler
         }
 
         /// <summary>
-        /// OnAsyncMethodEnd callback
+        /// OnMethodEnd callback
         /// </summary>
         /// <typeparam name="TTarget">Type of the target</typeparam>
-        /// <typeparam name="TResponse">Type of the response, in an async scenario will be T of Task of T</typeparam>
+        /// <typeparam name="TResponse">Type of the response</typeparam>
         /// <param name="instance">Instance value, aka `this` of the instrumented method.</param>
         /// <param name="responseMessage">HttpResponse message instance</param>
         /// <param name="exception">Exception instance in case the original code threw an exception.</param>
         /// <param name="state">Calltarget state value</param>
         /// <returns>A response value, in an async scenario will be T of Task of T</returns>
-        public static TResponse OnAsyncMethodEnd<TTarget, TResponse>(TTarget instance, TResponse responseMessage, Exception exception, CallTargetState state)
+        public static CallTargetReturn<TResponse> OnMethodEnd<TTarget, TResponse>(TTarget instance, TResponse responseMessage, Exception exception, CallTargetState state)
             where TResponse : IHttpResponseMessage
         {
-            return HttpClientHandlerCommon.OnMethodEnd(instance, responseMessage, exception, state);
+            var response = HttpClientHandlerCommon.OnMethodEnd(instance, responseMessage, exception, state);
+            return new CallTargetReturn<TResponse>(response);
         }
     }
 }
