+++ src/Datadog.Trace.ClrProfiler.Managed/AutoInstrumentation/Http/WinHttpHandler/IRequestHeaders.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 
-namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.SocketsHttpHandler
+namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.Http.WinHttpHandler
 {
     /// <summary>
     /// RequestHeaders interface for ducktyping
