+++ test/benchmarks/Benchmarks.Trace/HttpClientBenchmark.cs
@@ -5,7 +5,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using BenchmarkDotNet.Attributes;
 using Datadog.Trace;
-using Datadog.Trace.ClrProfiler.AutoInstrumentation.HttpClientHandler;
+using Datadog.Trace.ClrProfiler.AutoInstrumentation.Http.HttpClientHandler;
 using Datadog.Trace.ClrProfiler.CallTarget;
 using Datadog.Trace.ClrProfiler.Emit;
 using Datadog.Trace.ClrProfiler.Integrations;
